add yo-contract
